requirements.txt
Flask==0.11
Jinja2==2.8
gunicorn==19.6.0
apscheduler==3.3.1
python-dotenv==0.6.3